Fix decimal a enteros

+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/AppJob.cs
@@ -5,12 +5,12 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class AppJob
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
         public string Queue { get; set; } = null!;
         public string Payload { get; set; } = null!;
-        public decimal Attempts { get; set; }
-        public decimal? ReservedAt { get; set; }
-        public decimal AvailableAt { get; set; }
-        public decimal CreatedAt { get; set; }
+        public int Attempts { get; set; }
+        public int? ReservedAt { get; set; }
+        public int AvailableAt { get; set; }
+        public int CreatedAt { get; set; }
     }
 }
