+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/CusdyCatUnidadTipo.cs
@@ -5,9 +5,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class CusdyCatUnidadTipo
     {
-        public decimal UnidadTipoId { get; set; }
+        public int UnidadTipoId { get; set; }
         public string? UnidadTipoCod { get; set; }
         public string? UnidadTipoDsc { get; set; }
-        public decimal UnidadTipoOrgId { get; set; }
+        public int UnidadTipoOrgId { get; set; }
     }
 }
