+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatCustomersV.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatCustomersV
     {
-        public decimal? CustomerId { get; set; }
+        public int? CustomerId { get; set; }
         public string? CustomerNumber { get; set; }
         public string? CustomerName { get; set; }
     }
