+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatItem.cs
@@ -21,9 +21,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? DoorType { get; set; }
         public string? SerialStatus { get; set; }
         public string? ParentSerial { get; set; }
-        public decimal? CustomerId { get; set; }
-        public decimal? ContractId { get; set; }
-        public decimal? ContractOrgId { get; set; }
+        public int? CustomerId { get; set; }
+        public int? ContractId { get; set; }
+        public int? ContractOrgId { get; set; }
         public DateTime? CreationDate { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
