+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatLicense.cs
@@ -5,7 +5,7 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatLicense
     {
-        public decimal LicenseId { get; set; }
+        public int LicenseId { get; set; }
         public string? LicenseNumber { get; set; }
         public string? FirstName { get; set; }
         public string? MiddleName { get; set; }
@@ -17,9 +17,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
         public string ExternalOperator { get; set; } = null!;
     }
