+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatMaterial.cs
@@ -8,11 +8,11 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? ItemCode { get; set; }
         public string? UomCode { get; set; }
         public string? ItemDescription { get; set; }
-        public decimal? Amount { get; set; }
+        public int? Amount { get; set; }
         public string? Workforce { get; set; }
         public string? Miscellaneous { get; set; }
-        public decimal AssemblyItemId { get; set; }
-        public decimal OrganizationId { get; set; }
+        public int AssemblyItemId { get; set; }
+        public int OrganizationId { get; set; }
         public DateTime? CreationDate { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
