+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyCatMiscItem.cs
@@ -5,8 +5,8 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyCatMiscItem
     {
-        public decimal MiscellaneousId { get; set; }
-        public decimal ItemId { get; set; }
+        public int MiscellaneousId { get; set; }
+        public int ItemId { get; set; }
         public string? ItemCode { get; set; }
         public string? ItemDescription { get; set; }
         public string? ContextValue { get; set; }
@@ -15,9 +15,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
