+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyDaTransaction.cs
@@ -5,13 +5,13 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyDaTransaction
     {
-        public decimal TrxId { get; set; }
+        public int TrxId { get; set; }
         public string TrxType { get; set; } = null!;
         public DateTime TrxDate { get; set; }
-        public decimal DaDocumentId { get; set; }
-        public decimal DepartmentId { get; set; }
-        public decimal EmployeeId { get; set; }
-        public decimal TrxDays { get; set; }
+        public int DaDocumentId { get; set; }
+        public int DepartmentId { get; set; }
+        public int EmployeeId { get; set; }
+        public int TrxDays { get; set; }
         public string NoReturnFlag { get; set; } = null!;
         public string Motive { get; set; } = null!;
         public string? Comments { get; set; }
@@ -26,10 +26,10 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal OrgId { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int OrgId { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
