+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyDeActivation.cs
@@ -5,13 +5,13 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyDeActivation
     {
-        public decimal ActivationId { get; set; }
+        public int ActivationId { get; set; }
         public DateTime? ActivationDate { get; set; }
-        public decimal BranchId { get; set; }
+        public int BranchId { get; set; }
         public string SerialNumber { get; set; } = null!;
         public string? Vin { get; set; }
         public string? TrxType { get; set; }
-        public decimal? CustomerId { get; set; }
+        public int? CustomerId { get; set; }
         public string? CustomerName { get; set; }
         public string? ContractNumber { get; set; }
         public string? DealType { get; set; }
@@ -26,10 +26,10 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
-        public decimal? DeliveryId { get; set; }
+        public int? DeliveryId { get; set; }
     }
 }
