+++ TipMexico.DigitalYard.Domain.Entity/EntityFramework/XxdyDeAttachment.cs
@@ -5,9 +5,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
 {
     public partial class XxdyDeAttachment
     {
-        public decimal AttachId { get; set; }
-        public decimal HeaderId { get; set; }
-        public decimal? LineId { get; set; }
+        public int AttachId { get; set; }
+        public int HeaderId { get; set; }
+        public int? LineId { get; set; }
         public string AttachType { get; set; } = null!;
         public string FilePath { get; set; } = null!;
         public string? ContextValue { get; set; }
@@ -16,9 +16,9 @@ namespace TipMexico.DigitalYard.Domain.Entity.EntityFramework
         public string? CustomField3 { get; set; }
         public string? CustomField4 { get; set; }
         public string? CustomField5 { get; set; }
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
         public DateTime? CreationDate { get; set; }
-        public decimal? LastUpdatedBy { get; set; }
+        public int? LastUpdatedBy { get; set; }
         public DateTime? LastUpdateDate { get; set; }
     }
 }
